+++ Simulator PIC16F84/Instruktionen/CALL.cs
@@ -26,7 +26,7 @@ namespace Simulator_PIC16F84.Instruktionen
     {
         
 
-        public CALL( byte k, ProgramCounter PC, Stack Stack, RegisterFileMap Reg ) : base(Reg)
+        public CALL( int k, ProgramCounter PC, Stack Stack, RegisterFileMap Reg ) : base(Reg)
         {
             this.k = k;
